backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
joblib==1.4.2
xgboost
mrmr_selection
